# ms2-orders/qa/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Impochina - ms2-orders",
    "description": "Microservicio de órdenes (creación, consulta y salud).",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo Impochina",
      "email": "soporte@impochina.local",
      "url": "http://localhost:8000/docs"
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "description": "Verifica el estado del servicio (ping).",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/orders/me": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "List My Orders",
        "description": "Lista las órdenes del usuario autenticado.",
        "operationId": "list_my_orders_orders_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderOut"
                  },
                  "title": "Response List My Orders Orders Me Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Create Order",
        "description": "Crea una nueva orden para el usuario autenticado.",
        "operationId": "create_order_orders_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OrderIn": {
        "properties": {
          "item": {
            "type": "string",
            "title": "Item"
          },
          "qty": {
            "type": "integer",
            "title": "Qty"
          }
        },
        "type": "object",
        "required": [
          "item",
          "qty"
        ],
        "title": "OrderIn"
      },
      "OrderOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "item": {
            "type": "string",
            "title": "Item"
          },
          "qty": {
            "type": "integer",
            "title": "Qty"
          }
        },
        "type": "object",
        "required": [
          "id",
          "item",
          "qty"
        ],
        "title": "OrderOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "tags": [
    {
      "name": "health",
      "description": "Endpoints de salud/monitor."
    },
    {
      "name": "orders",
      "description": "Gestión de órdenes del usuario."
    }
  ]
}